Project Finished: Ready For Review
I got everything I needed, all in one day.

My first line of duty was to link all the needed stylesheets (just one here), then create a video tag and insert two files to play within the page. CSS was naturally then added ontop of that, but a .vtt file was created and managed to make sure captions then could be inserted. The captions so far only worked on Safari, but will hopefully work on Chrome when the website is published.

Oh yeah, and documentation was also completed as well; including the interesting new feat of doing "NOTE" within the .vtt file to not disturb the code.